% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCropPlotter.R
\name{aquacropplotter}
\alias{aquacropplotter}
\title{Launching the app}
\usage{
aquacropplotter()
}
\value{
launching the app
}
\description{
Load and visualize AquaCrop results
}
\examples{
aquacropplotter()

}
